requirements.txt
requests==2.31.0
Jinja2==3.1.2
python-dateutil==2.8.2
pytz==2023.3 